addons/mealie-import-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mealie-import-orchestrator"
version = "0.1.0"
description = "External addon orchestrating the canonical Mealie import workflow"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=0.20.0",
    "requests>=2.32.0",
]

[project.scripts]
mealie-import-orchestrator = "mealie_import_orchestrator.cli:main"
smoke-test = "mealie_import_orchestrator.smoke_test:main"

[tool.setuptools.package-data]
mealie_import_orchestrator = []

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
